deploy/requirements.txt
absl-py
jsonschema
pyyaml
ruamel.yaml
